Create parse_csv.py
functions dedicated to parsing through csv files and acquiring only longitude and latitude data. The function also utilizes a clean coordinate function which removes unnecessary characters like the degree symbol or "N,S,E,W". The clean function also checks to see whether the coordinates are in dms format and if so converts to decimal values.